test: modify how the `max_character_count_layer_filter` adjusts layers
This is a test of the performance impact of the
`max_character_count_layer_filter` when its behavior is changed.

Background
==========

This filter is used to exclude results from the address layer when an
autocomplete query consists of less than a configured number of
characters (usually set to 1 or 2). This helps performance for queries
that otherwise match millions of documents since the input is so short.

It currently works by adding an additional filter to the generated
Elasticsearch query. This filter enumerates _all layers_ except the
address layer and tells Elasticsearch to require documents to match
those layers. This is the most performant way to exclude a single layer
(Elasticsearch does not support fast `not`-style queries).

However, because there is often _another_ filter that limits results to
certain layers (this one coming from the `layers` parameter), it seems
redundant to create a second filter and force Elasticsearch to figure
out how to manage both of them.

Potential impact
================

Elasticsearch is very good at making hard filters, like the `terms`
query we use for managing layers. That said, its possible that reducing
two `terms` queries to a single query that has the same effect will be a
little faster.

It might be that most of the practical performance impact even comes
from the reduction in the size of the query. Its that many fewer bytes
to transfer, that much less JSON to serialize/deserialize, etc.

Even if there's no performance impact, it might be nice to avoid having
two `terms` queries that look very similar. I think I recall people
being confused by this in the past.

How this change works
=====================

This change is just a test of the general concept of modifying the
existing `terms` query that comes from the `layers` parameter to exclude
the `address` layer whenever the input length determines that's a good
idea.

It does to by using the existing code to modify the query variable,
instead of adding a new clause to the query.

This is probably not how we want to manage things at the end of the day,
but it's a quick way to see if this is faster in any way.

Why now?
========

With the negative sources and layers work from
https://github.com/pelias/pelias/issues/844, we're doing some overhaul
of how we manage these filters. Its a good time to look at things that
are either inefficient or confusing.

In particular, because we _also_ have the [address layer
sanitizer](https://github.com/pelias/api/blob/master/sanitizer/_address_layer_filter.js),
which similarly excludes the `address` layer under certain conditions,
it would make sense to structure the two bits of functionality in a
single way. Having one modify the Elasticsearch query, and the other
modify the set of `layers` sent to the query, seems a bit complex.
Standardizing on one or the other would be helpful in keeping things
clean.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch test-autocomplete-address-exclude-changes
# Changes to be committed:
#	modified:   query/view/max_character_count_layer_filter.js
#	modified:   test/unit/query/view/max_character_count_layer_filter.js
#